html/canvas/offscreen/canvas-host/2d.canvas.host.size.attributes.removed.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.canvas.host.size.attributes.removed</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.host.size.attributes.removed</h1>
<p class="desc">Removing content attributes reverts to default size</p>


<script>
var t = async_test("Removing content attributes reverts to default size");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  _assertSame(canvas.width, 120, "canvas.width", "120");
  canvas.removeAttribute('width');
  _assertSame(canvas.width, 300, "canvas.width", "300");
  t.done();

});
</script>
